interview-agent/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0